console/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
aiosqlite>=0.19
# Optional: needed when VIOLAWAKE_DB_URL points to PostgreSQL, such as Railway.
asyncpg>=0.29
python-jose[cryptography]>=3.3
bcrypt>=4.0
python-multipart>=0.0.9
sse-starlette>=1.6
pydantic>=2.0
pydantic-settings>=2.0
email-validator>=2.0
httpx>=0.27
numpy>=1.24
scipy>=1.11
stripe>=8.0.0
sentry-sdk[fastapi]>=2.0
tomli>=2.0; python_version < "3.11"
boto3>=1.34
